create NoteGrid component
